do not convert

+++ ubana/searchingfornues/Selection/AnalysisTools/ContainmentAnalysis_tool.cc
@@ -301,8 +301,8 @@ void ContainmentAnalysis::resetTTree(TTree *_tree)
 {
 
   // implicit conversion to float reported by c14
-  _dvtx = (float)std::numeric_limits<int>::max();
-  _dtrk = (float)std::numeric_limits<int>::max();
+  _dvtx = std::numeric_limits<int>::max();
+  _dtrk = std::numeric_limits<int>::max();
 
   return;
 }
